requirements.txt
python-dotenv==0.17.1
requests==2.26.0